Fixed Player damage bug
Now you don't die when attacking...
Only took about two days.
Instead of looking for the swordcollider gameobject I now check is the attack animation is active.
A bug will persist in which the player can't be damaged at all when attacking(say if player was attacked from the back). That will need fixing.
For now I will start working on Enemy movement(trackers and AI).